Created MVC package structure.
